requirements.txt
fastapi 
uvicorn
joblib
pandas
xgboost
requests
PyPDF2
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
bcrypt==4.0.1
pydantic[email]
